+++ src/games/valve.rs
@@ -35,6 +35,17 @@ game_query_mod!(
     27004
 );
 game_query_mod!(
+    conanexiles,
+    "Conan Exiles",
+    Engine::new(440_900),
+    27015,
+    GatheringSettings {
+        players: false,
+        rules: true,
+        check_app_id: true,
+    }
+);
+game_query_mod!(
     counterstrike,
     "Counter-Strike",
     Engine::new_gold_src(false),
@@ -107,6 +118,7 @@ game_query_mod!(
     27015
 );
 game_query_mod!(theforest, "The Forest", Engine::new(556_450), 27016);
+game_query_mod!(thefront, "The Front", Engine::new(2_285_150), 27015);
 game_query_mod!(unturned, "Unturned", Engine::new(304_930), 27015);
 game_query_mod!(
     valheim,